Polymorphism with Abstract Class and Interface Doctor and Nurse are concrete classes that extend MedicalStaff. Both override the abstract method performDuties. Doctor implements Diagnosable, MedicationProvider, and Treatable, providing specific implementations for the methods. Patient implements Payable. Administrator implements Manageable.
Final Class, Method, and Variable
The Patient class is marked as final.
The method printPatientInfo() in the Patient class is marked as final.
The variable role in the Doctor class is marked as final.

Static Block, Method, and Variable
In HospitalUtils:
Static Block: initializes static variables and performing any other setup required when the class is first loaded.
Static variables: totalPatients and departmentPatientCount.
Static Methods: registerPatient(String department), showTotalPatients(), showPatientCountByDepartment().